+++ docs/root/intro/version_history.rst
@@ -96,6 +96,7 @@ Version history
 * stats: added unit support to histogram.
 * tcp_proxy: the default :ref:`idle_timeout
   <envoy_api_field_config.filter.network.tcp_proxy.v2.TcpProxy.idle_timeout>` is now 1 hour.
+* tcp_proxy: added :ref:`hash_with_source_ip<envoy_api_field_config.filter.network.tcp_proxy.v2.TcpProxy.hash_with_source_ip>`
 * thrift_proxy: fix crashing bug on invalid transport/protocol framing
 * thrift_proxy: add support for stripping service name from method when using the multiplexed protocol.
 * tls: added verification of IP address SAN fields in certificates against configured SANs in the